requirements.txt
django
djangorestframework
django-cors-headers
drf-yasg
django-rest-swagger